server/src/requirements.txt
Flask
Flask-PyMongo
flask-swagger
swagger-ui-py
requests